Inline here_doc::open function
Refactoring...
