+++ yash-semantics/src/redir/here_doc.rs
@@ -16,18 +16,12 @@
 
 //! Here-documents
 
-use super::Error;
 use super::ErrorCause;
-use super::FdSpec;
-use crate::expansion::expand_text;
 use std::path::Path;
 use yash_env::io::Fd;
-use yash_env::semantics::ExitStatus;
 use yash_env::system::Errno;
 use yash_env::Env;
 use yash_env::System;
-use yash_syntax::source::Location;
-use yash_syntax::syntax::HereDoc;
 
 async fn fill_content(env: &mut Env, fd: Fd, content: &str) -> Result<(), Errno> {
     env.system.write_all(fd, content.as_bytes()).await?;
@@ -55,35 +49,11 @@ pub(super) async fn open_fd(env: &mut Env, content: String) -> Result<Fd, ErrorC
     }
 }
 
-/// Opens a here-document.
-///
-/// This function expands the here-document content to an anonymous temporary
-/// file and returns a file descriptor to the file you can read the content
-/// from.
-#[allow(clippy::await_holding_refcell_ref)]
-pub(super) async fn open(
-    env: &mut Env,
-    here_doc: &HereDoc,
-) -> Result<(FdSpec, Location, Option<ExitStatus>), Error> {
-    let (content, exit_status) = expand_text(env, &here_doc.content.borrow()).await?;
-    let location = here_doc.delimiter.location.clone();
-    match open_fd(env, content).await {
-        Ok(fd) => Ok((FdSpec::Owned(fd), location, exit_status)),
-        Err(cause) => Err(Error { cause, location }),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tests::echo_builtin;
-    use crate::tests::in_virtual_system;
-    use crate::tests::return_builtin;
-    use assert_matches::assert_matches;
     use futures_util::FutureExt;
-    use std::cell::RefCell;
     use yash_env::System;
-    use yash_syntax::syntax::Text;
 
     #[test]
     fn open_fd_and_read_from_it() {
@@ -97,48 +67,5 @@ mod tests {
         let mut buffer = [0; 30];
         let count = env.system.read(fd, &mut buffer).unwrap();
         assert_eq!(std::str::from_utf8(&buffer[..count]), Ok(text));
-    }
-
-    #[test]
-    fn empty_here_doc() {
-        let mut env = Env::new_virtual();
-        let here_doc = HereDoc {
-            delimiter: "END".parse().unwrap(),
-            remove_tabs: false,
-            content: RefCell::new(Text(vec![])),
-        };
-
-        let (fd_spec, location, exit_status) =
-            open(&mut env, &here_doc).now_or_never().unwrap().unwrap();
-        assert_matches!(fd_spec, FdSpec::Owned(fd) => {
-            let mut buffer = [0; 1];
-            let read_count = env.system.read(fd, &mut buffer).unwrap();
-            assert_eq!(read_count, 0);
-        });
-        assert_eq!(location, here_doc.delimiter.location);
-        assert_eq!(exit_status, None);
-    }
-
-    #[test]
-    fn here_doc_with_command_substitution() {
-        in_virtual_system(|mut env, _pid, _state| async move {
-            env.builtins.insert("echo", echo_builtin());
-            env.builtins.insert("return", return_builtin());
-            let here_doc = HereDoc {
-                delimiter: "EOF".parse().unwrap(),
-                remove_tabs: false,
-                content: RefCell::new("$(echo foo)$(echo bar; return -n 42)\n".parse().unwrap()),
-            };
-
-            let (fd_spec, location, exit_status) = open(&mut env, &here_doc).await.unwrap();
-            assert_matches!(fd_spec, FdSpec::Owned(fd) => {
-                let mut buffer = [0; 8];
-                let read_count = env.system.read(fd, &mut buffer).unwrap();
-                assert_eq!(read_count, 7);
-                assert_eq!(&buffer[..7], b"foobar\n");
-            });
-            assert_eq!(location, here_doc.delimiter.location);
-            assert_eq!(exit_status, Some(ExitStatus(42)));
-        })
     }
 }
